Bring back color mixin helpers from compat to main.
